+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Only show links to views requiring edit permission if the user has that permission.
 - Add a get_absolute_url method for `WorkspaceData` which returns the absolute url of the associated workspace.
 - Add a required "workspace_detail_template_name" field for workspace adapters.
+- Add Workspace and ManagedGroup methods to check whether a group has access to a workspace.
 
 ## 0.3 (2022-09-27)